Remove redux from files
