--- src/actions/index.js
@@ -1,3 +1,0 @@
-export const setUrls = () => ({
-  type: 'SET_URLS'
-});
