+++ src/components/App/App.js
@@ -1,0 +1,32 @@
+import React, { Component } from 'react';
+import './App.css';
+import { getUrls } from '../../apiCalls';
+import UrlContainer from '../UrlContainer/UrlContainer';
+import UrlForm from '../UrlForm/UrlForm';
+
+export class App extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      urls: []
+    }
+  }
+
+  componentDidMount() {
+  }
+
+  render() {
+    return (
+      <main className="App">
+        <header>
+          <h1>URL Shortener</h1>
+          <UrlForm />
+        </header>
+
+        <UrlContainer urls={this.state.urls}/>
+      </main>
+    );
+  }
+}
+
+export default App;
