--- src/reducers/index.js
@@ -1,8 +1,0 @@
-import { combineReducers } from 'redux';
-import { urls } from './urls';
-
-const rootReducer = combineReducers({
-  urls,
-});
-
-export default rootReducer;
