--- src/reducers/urls.js
@@ -1,8 +1,0 @@
-export const urls = (state = [], action) => {
-  switch (action.type) {
-    case 'SET_URLS':
-      return action.urls;
-    default:
-      return state;
-  }
-};
